requirements.txt
ortools>=9.4.1874
python-dotenv>=1.0.0
supabase>=2.0.0
python-dateutil>=2.8.2 